+++ app/overrides/wiki/show.rb
@@ -1,5 +1,5 @@
 Deface::Override.new virtual_path: 'wiki/show',
                      name: 'addto-wiki-show',
                      insert_before: 'div.contextual',
-                     original: 'add94a807567d348a276d12ad25929e5e9bb03d6',
+                     original: '6b0cb1646d5e2cb23feee1805949e266036581e6',
                      partial: 'wiki/show_additionals'