+++ init.rb
@@ -4,7 +4,7 @@ Redmine::Plugin.register :additionals do
   name 'Additionals'
   author 'AlphaNodes GmbH'
   description 'Customizing Redmine, providing wiki macros and act as a library/function provider for other Redmine plugins'
-  version '2.0.17'
+  version '2.0.18'
   author_url 'https://alphanodes.com/'
   url 'https://github.com/alphanodes/additionals'